examples/integrations/crewai/requirements.txt
# AxonFlow SDK
axonflow>=0.2.0

# CrewAI
crewai>=0.5.0
crewai-tools>=0.4.0

# LLM Provider
langchain-openai>=0.0.5

# Environment management
python-dotenv>=1.0.0
